Migrate NextAuth from v4 to v5 beta with configuration updates
