+++ package.json
@@ -59,6 +59,7 @@
     "embla-carousel-autoplay": "^8.3.0",
     "embla-carousel-react": "^8.3.0",
     "formidable": "^3.5.1",
+    "framer-motion": "^12.22.0",
     "get-audio-duration": "^4.0.1",
     "html-react-parser": "^5.1.18",
     "input-otp": "^1.2.4",
@@ -66,7 +67,8 @@
     "jsonwebtoken": "^9.0.2",
     "knex": "^3.1.0",
     "lucide-react": "^0.439.0",
-    "next-auth": "^4.24.7",
+    "next": "^15.3.4",
+    "next-auth": "^5.0.0-beta.29",
     "next-themes": "^0.3.0",
     "pdf-parse": "^1.1.1",
     "pdfjs-dist": "^5.3.31",
@@ -86,8 +88,6 @@
     "react-select": "^5.8.1",
     "react-select-async-paginate": "^0.7.6",
     "recharts": "^2.12.7",
-    "framer-motion": "^12.22.0",
-    "next": "^15.3.4",
     "sanitize-html": "^2.13.1",
     "sharp": "^0.33.5",
     "sonner": "^1.5.0",
